Moved physics code to game_physics.c
