+++ DragonQuestinoWinDev/DragonQuestinoWinDev/DragonQuestinoWinDev.vcxproj
@@ -42,6 +42,7 @@
     <ClCompile Include="..\..\DragonQuestino\game_data.c" />
     <ClCompile Include="..\..\DragonQuestino\game_input.c" />
     <ClCompile Include="..\..\DragonQuestino\game_items.c" />
+    <ClCompile Include="..\..\DragonQuestino\game_physics.c" />
     <ClCompile Include="..\..\DragonQuestino\game_render.c" />
     <ClCompile Include="..\..\DragonQuestino\input.c" />
     <ClCompile Include="..\..\DragonQuestino\math.c" />
